housekeeping: refactor tools screen into separate files
